major release bora

+++ lib/telas/home/gatos/forum/edit/delete_post.dart
@@ -19,14 +19,21 @@ class _DeletePostState extends State<DeletePost> {
   _deletar(int postN) async {
     FirebaseDatabase database = FirebaseDatabase.instance;
     DatabaseReference ref = database.ref("posts/$postN");
+    await ref.remove();
+    if (!mounted) return;
+    Flushbar(
+      message: "Excluído com sucesso!",
+      duration: const Duration(seconds: 3),
+      margin: const EdgeInsets.all(20),
+      borderRadius: BorderRadius.circular(50),
+    ).show(context);
     if (snapshotForum!.child("$postN/img").value != null) {
       Reference refI = FirebaseStorage.instance.ref("posts/$postN.webp");
       Uint8List? lista = await refI.getData();
       if (lista?.isNotEmpty ?? false) {
-        refI.delete();
+        await refI.delete();
       }
     }
-    ref.remove();
   }
 
   @override
@@ -44,26 +51,14 @@ class _DeletePostState extends State<DeletePost> {
       actions: [
         TextButton(
           onPressed: () => Navigator.pop(context),
-          child: const Text(
-            "CANCELAR",
-          ),
+          child: const Text("CANCELAR"),
         ),
         ElevatedButton(
           onPressed: () {
             _deletar(widget.post);
             Navigator.pop(context);
-            Flushbar(
-              message: "Excluído com sucesso!",
-              duration: const Duration(seconds: 3),
-              margin: const EdgeInsets.all(20),
-              borderRadius: BorderRadius.circular(50),
-            ).show(
-              context,
-            );
           },
-          child: const Text(
-            "OK",
-          ),
+          child: const Text("OK"),
         ),
       ],
     );
